requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
python-dotenv>=0.19
Flask-Migrate>=3.0 
# psycopg2-binary # Add if PostgreSQL is intended, for now SQLite
